+++ internal/controller/operator/factory/vmauth/vmusers_config_test.go
@@ -20,6 +20,198 @@ import (
 )
 
 func Test_genUserCfg(t *testing.T) {
+
+	type opts struct {
+		user              *vmv1beta1.VMUser
+		crdURLCache       map[string]string
+		predefinedObjects []runtime.Object
+	}
+	f := func(wantConfig string, opts opts) {
+		t.Helper()
+		cr := &vmv1beta1.VMAuth{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "test-auth",
+				Namespace: "default",
+			},
+		}
+		ctx := context.TODO()
+		fclient := k8stools.GetTestClientWithObjects(opts.predefinedObjects)
+		ac := getAssetsCache(ctx, fclient, cr)
+		got, err := genUserCfg(opts.user, opts.crdURLCache, cr, ac)
+		if err != nil {
+			t.Fatalf("unexpected genUserCfg() error = %s", err)
+		}
+		szd, err := yaml.Marshal(got)
+		if err != nil {
+			t.Fatalf("BUG: cannot serialize result: %v", err)
+		}
+		assert.Equal(t, wantConfig, string(szd))
+	}
+
+	// with vmcluster crd and specific tenant suffix
+	o := opts{
+		user: &vmv1beta1.VMUser{
+			Spec: vmv1beta1.VMUserSpec{
+				Name:        ptr.To("user1"),
+				BearerToken: ptr.To("secret-token"),
+				TargetRefs: []vmv1beta1.TargetRef{
+					{
+						CRD: &vmv1beta1.CRDRef{
+							Kind:      "VMCluster/vminsert",
+							Name:      "vminsert",
+							Namespace: "monitoring",
+						},
+						TargetPathSuffix: "/insert/1",
+					},
+					{
+						CRD: &vmv1beta1.CRDRef{
+							Kind:      "VMCluster/vmselect",
+							Namespace: "monitoring",
+							Name:      "vmselect",
+						},
+						TargetPathSuffix: "/select/1",
+					},
+				},
+			},
+		},
+		crdURLCache: map[string]string{
+			"VMCluster/vminsert/monitoring/vminsert": "http://vminsert.monitoring.svc:8481",
+			"VMCluster/vmselect/monitoring/vmselect": "http://vmselect.monitoring.svc:8482",
+		},
+	}
+	wantConfig := `url_map:
+- url_prefix:
+  - http://vminsert.monitoring.svc:8481/insert/1
+  src_paths:
+  - /newrelic/.*
+  - /opentelemetry/.*
+  - /prometheus/api/v1/write
+  - /prometheus/api/v1/import.*
+  - /influx/.*
+  - /datadog/.*
+- url_prefix:
+  - http://vmselect.monitoring.svc:8482/select/1
+  src_paths:
+  - /vmui.*
+  - /vmui.*
+  - /graph.*
+  - /prometheus/graph.*
+  - /prometheus/vmui.*
+  - /prometheus/api/v1/label.*
+  - /prometheus/api/v1/query.*
+  - /prometheus/api/v1/rules
+  - /prometheus/api/v1/alerts
+  - /prometheus/api/v1/metadata
+  - /prometheus/api/v1/series.*
+  - /prometheus/api/v1/status.*
+  - /prometheus/api/v1/export.*
+  - /prometheus/federate
+  - /admin/tenants
+  - /api/v1/status/.*
+  - /api/v1/rules
+  - /internal/resetRollupResultCache
+  - /prometheus/api/v1/admin/.*
+  - /prometheus.*-debug
+  - /prometheus/prettify-query
+  - /prometheus/api/v1/notifiers
+  - /prometheus/api/v1/query_exemplars
+name: user1
+bearer_token: secret-token
+`
+	f(wantConfig, o)
+
+	// with cluster CRD and empty targetSuffix
+	o = opts{
+		user: &vmv1beta1.VMUser{
+			Spec: vmv1beta1.VMUserSpec{
+				Name:        ptr.To("user1"),
+				BearerToken: ptr.To("secret-token"),
+				TargetRefs: []vmv1beta1.TargetRef{
+					{
+						CRD: &vmv1beta1.CRDRef{
+							Kind:      "VMCluster/vminsert",
+							Name:      "vminsert",
+							Namespace: "monitoring",
+						},
+					},
+					{
+						CRD: &vmv1beta1.CRDRef{
+							Kind:      "VMCluster/vmselect",
+							Namespace: "monitoring",
+							Name:      "vmselect",
+						},
+					},
+				},
+			},
+		},
+		crdURLCache: map[string]string{
+			"VMCluster/vminsert/monitoring/vminsert": "http://vminsert.monitoring.svc:8481",
+			"VMCluster/vmselect/monitoring/vmselect": "http://vmselect.monitoring.svc:8482",
+		},
+	}
+	wantConfig = `url_map:
+- url_prefix:
+  - http://vminsert.monitoring.svc:8481
+  src_paths:
+  - /insert/.*
+- url_prefix:
+  - http://vmselect.monitoring.svc:8482
+  src_paths:
+  - /select/.*
+  - /admin/.*
+name: user1
+bearer_token: secret-token
+`
+	f(wantConfig, o)
+
+	// with cluster CRD and mixed targetSuffix
+	o = opts{
+		user: &vmv1beta1.VMUser{
+			Spec: vmv1beta1.VMUserSpec{
+				Name:        ptr.To("user1"),
+				BearerToken: ptr.To("secret-token"),
+				TargetRefs: []vmv1beta1.TargetRef{
+					{
+						CRD: &vmv1beta1.CRDRef{
+							Kind:      "VMCluster/vminsert",
+							Name:      "vminsert",
+							Namespace: "monitoring",
+						},
+						Paths:            []string{"/"},
+						TargetPathSuffix: "/insert/1",
+					},
+					{
+						CRD: &vmv1beta1.CRDRef{
+							Kind:      "VMCluster/vmselect",
+							Namespace: "monitoring",
+							Name:      "vmselect",
+						},
+					},
+				},
+			},
+		},
+		crdURLCache: map[string]string{
+			"VMCluster/vminsert/monitoring/vminsert": "http://vminsert.monitoring.svc:8481",
+			"VMCluster/vmselect/monitoring/vmselect": "http://vmselect.monitoring.svc:8482",
+		},
+	}
+	wantConfig = `url_map:
+- url_prefix:
+  - http://vminsert.monitoring.svc:8481/insert/1
+  src_paths:
+  - /.*
+- url_prefix:
+  - http://vmselect.monitoring.svc:8482
+  src_paths:
+  - /select/.*
+  - /admin/.*
+name: user1
+bearer_token: secret-token
+`
+
+	f(wantConfig, o)
+
+	// TODO: rewrite other tests with f-test
 	type args struct {
 		user        *vmv1beta1.VMUser
 		crdURLCache map[string]string
@@ -1850,38 +2042,14 @@ unauthorized_user:
   - url_prefix:
     - http://vmselect-main-cluster.default.svc:8481
     src_paths:
-    - /vmui.*
-    - /vmui/vmui
-    - /graph
-    - /prometheus/graph
-    - /prometheus/vmui.*
-    - /prometheus/api/v1/label.*
-    - /graphite.*
-    - /prometheus/api/v1/query.*
-    - /prometheus/api/v1/rules
-    - /prometheus/api/v1/alerts
-    - /prometheus/api/v1/metadata
-    - /prometheus/api/v1/rules
-    - /prometheus/api/v1/series.*
-    - /prometheus/api/v1/status.*
-    - /prometheus/api/v1/export.*
-    - /prometheus/federate
-    - /prometheus/api/v1/admin/tsdb/delete_series
-    - /admin/tenants
-    - /api/v1/status/.*
-    - /internal/resetRollupResultCache
-    - /prometheus/api/v1/admin/.*
+    - /select/.*
+    - /admin/.*
     headers:
     - 'Authorization: Basic c29tZS0xOnNvbWUtMg=='
   - url_prefix:
     - http://vminsert-main-cluster.default.svc:8480
     src_paths:
-    - /newrelic/.*
-    - /opentelemetry/.*
-    - /prometheus/api/v1/write
-    - /prometheus/api/v1/import.*
-    - /influx/.*
-    - /datadog/.*
+    - /insert/.*
     headers:
     - 'Authorization: Basic c29tZS0xOnNvbWUtMg=='
   name: user1